requirements.txt
# Runtime dependencies
requests>=2.28.1
prettytable>=3.6.0
python-dotenv>=0.21.0
colorama>=0.4.6
html2text>=2020.1.16
tk>=0.1.0
pyperclip>=1.8.2

# Development dependencies
pytest>=7.3.1
flake8>=6.0.0
mypy>=1.3.0
black>=23.3.0
isort>=5.12.0